+++ internal/inspect/dependencies/pydeps/testdata/lib/utils.py
@@ -1,4 +1,0 @@
-from somelib import somefunc
-
-def foo():
-    somefunc()